refactor: change to follow the GitLab's testing best practices
see https://docs.gitlab.com/ee/development/testing_guide/best_practices.html
