--- spec/helpers.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require "json"
-
-module Helpers
-  def io_to_queue(io)
-    io.rewind
-    io.readlines.map { |line| JSON.parse line }
-  end
-end
